+++ src/app/(main)/home/page.tsx
@@ -1,0 +1,7 @@
+import Home from "@/components/modules/Home/Main";
+
+function Page() {
+  return <Home />;
+}
+
+export default Page;
