--- src/app/(main)/homeguest/page.tsx
@@ -1,7 +1,0 @@
-import MainGuest from "@/components/modules/homeguest/Main";
-
-function Page() {
-  return <MainGuest />;
-}
-
-export default Page;
