--- src/app/(main)/homeuser/page.tsx
@@ -1,7 +1,0 @@
-import HomeUser from "@/components/modules/homeuser/Main";
-
-function Page() {
-  return <HomeUser />;
-}
-
-export default Page;
