+++ src/components/modules/Home/Main/styles.ts
@@ -1,14 +1,18 @@
+import { Flex } from "antd";
 import styled from "styled-components";
 
-
-export const MainWrapper = styled.div`
-  display: flex;
+export const HomeWrapper = styled(Flex)`
+  flex-direction: row;
   justify-content: center;
+  width: 100%;
+  position: relative;
+`;
+
+export const MainWrapper = styled(Flex)`
   margin-top: 40px;
 `;
 
-export const ContentWrapper = styled.div`
-  display: flex;
+export const ContentWrapper = styled(Flex)`
   flex-direction: column;
   align-items: center;
   flex: 1;
@@ -17,25 +21,25 @@ export const ContentWrapper = styled.div`
 export const PostContainer = styled.div`
   max-width: 800px;
   width: 100%;
-  justify-content: center;
 `;
 
-export const TagsContainer = styled.div`
-  display: flex;
+export const TagsContainer = styled(Flex)`
   justify-content: center;
-  gap:10px;
-  margin-bottom:20px;
+  margin-bottom: 20px;
+
+  gap: 10px;
 `;
 
 export const Tag = styled.span`
   background-color: transparent;
-  border: 1px solid #B9B4C7;
-  border-radius: 4px;a
+  color: #b9b4c7;
+
+  border: 1px solid #b9b4c7;
+  border-radius: 4px;
+
   padding: 2px 8px;
-  color: #B9B4C7;
   cursor: pointer;
   &:hover {
     background-color: #ccc;
   }
 `;
-
